requirements.txt
PyQt6>=6.6.0
pywin32>=306
Pillow>=10.0.0
pytesseract>=0.3.10
